fix: move files to lib
