+++ lib/git-host-info.js
