+++ lib/git-host.js
