+++ package.json
@@ -2,7 +2,7 @@
   "name": "hosted-git-info",
   "version": "4.1.0",
   "description": "Provides metadata and conversions from repository urls for GitHub, Bitbucket and GitLab",
-  "main": "index.js",
+  "main": "./lib/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/npm/hosted-git-info.git"
