+++ test/bitbucket.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 'use strict'
-const HostedGit = require('../index')
+const HostedGit = require('..')
 const t = require('tap')
 
 const invalid = [
@@ -14,7 +14,7 @@ const invalid = [
 
 // assigning the constructor here is hacky, but the only way to make assertions that compare
 // a subset of properties to a found object pass as you would expect
-const GitHost = require('../git-host')
+const GitHost = require('../lib/git-host')
 const defaults = { constructor: GitHost, type: 'bitbucket', user: 'foo', project: 'bar' }
 
 const valid = {
