+++ test/github.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-const HostedGit = require('../index')
+const HostedGit = require('..')
 const t = require('tap')
 
 const invalid = [
@@ -24,7 +24,7 @@ const invalid = [
 
 // assigning the constructor here is hacky, but the only way to make assertions that compare
 // a subset of properties to a found object pass as you would expect
-const GitHost = require('../git-host')
+const GitHost = require('../lib/git-host')
 const defaults = { constructor: GitHost, type: 'github', user: 'foo', project: 'bar' }
 const valid = {
   // extreme shorthand
